+++ examples/custom_config.rs
@@ -1,5 +1,6 @@
 use cpal::traits::HostTrait;
 use cpal::{BufferSize, SampleFormat};
+use rodio::effects::amplify::Factor;
 use rodio::source::SineWave;
 use rodio::Source;
 use std::error::Error;
@@ -25,7 +26,7 @@ fn main() -> Result<(), Box<dyn Error>> {
     let mixer = stream_handle.mixer();
 
     let wave = SineWave::new(740.0)
-        .amplify(0.1)
+        .amplify(Factor::Linear(0.1))
         .take_duration(Duration::from_secs(1));
     mixer.add(wave);