+++ examples/signal_generator.rs
@@ -3,6 +3,8 @@
 use std::error::Error;
 use std::num::NonZero;
 
+use rodio::effects::amplify::Factor;
+
 fn main() -> Result<(), Box<dyn Error>> {
     use rodio::source::{chirp, Function, SignalGenerator, Source};
     use std::thread;
@@ -17,7 +19,7 @@ fn main() -> Result<(), Box<dyn Error>> {
     println!("Playing 1000 Hz tone");
     stream_handle.mixer().add(
         SignalGenerator::new(sample_rate, 1000.0, Function::Sine)
-            .amplify(0.1)
+            .amplify(Factor::Linear(0.1))
             .take_duration(test_signal_duration),
     );
 
@@ -26,7 +28,7 @@ fn main() -> Result<(), Box<dyn Error>> {
     println!("Playing 10,000 Hz tone");
     stream_handle.mixer().add(
         SignalGenerator::new(sample_rate, 10000.0, Function::Sine)
-            .amplify(0.1)
+            .amplify(Factor::Linear(0.1))
             .take_duration(test_signal_duration),
     );
 
@@ -35,7 +37,7 @@ fn main() -> Result<(), Box<dyn Error>> {
     println!("Playing 440 Hz Triangle Wave");
     stream_handle.mixer().add(
         SignalGenerator::new(sample_rate, 440.0, Function::Triangle)
-            .amplify(0.1)
+            .amplify(Factor::Linear(0.1))
             .take_duration(test_signal_duration),
     );
 
@@ -44,7 +46,7 @@ fn main() -> Result<(), Box<dyn Error>> {
     println!("Playing 440 Hz Sawtooth Wave");
     stream_handle.mixer().add(
         SignalGenerator::new(sample_rate, 440.0, Function::Sawtooth)
-            .amplify(0.1)
+            .amplify(Factor::Linear(0.1))
             .take_duration(test_signal_duration),
     );
 
@@ -53,7 +55,7 @@ fn main() -> Result<(), Box<dyn Error>> {
     println!("Playing 440 Hz Square Wave");
     stream_handle.mixer().add(
         SignalGenerator::new(sample_rate, 440.0, Function::Square)
-            .amplify(0.1)
+            .amplify(Factor::Linear(0.1))
             .take_duration(test_signal_duration),
     );
 
@@ -62,7 +64,7 @@ fn main() -> Result<(), Box<dyn Error>> {
     println!("Playing 20-10000 Hz Sweep");
     stream_handle.mixer().add(
         chirp(sample_rate, 20.0, 10000.0, Duration::from_secs(1))
-            .amplify(0.1)
+            .amplify(Factor::Linear(0.1))
             .take_duration(test_signal_duration),
     );