+++ src/lib.rs
@@ -70,6 +70,7 @@
 #![cfg_attr(feature = "playback", doc = "```no_run")]
 //! use std::time::Duration;
 //! use rodio::{MixerDeviceSink, Player};
+//! use rodio::effects::amplify::Factor;
 //! use rodio::source::{SineWave, Source};
 //!
 //! // _stream must live as long as the sink
@@ -78,7 +79,9 @@
 //! let player = rodio::Player::connect_new(&handle.mixer());
 //!
 //! // Add a dummy source of the sake of the example.
-//! let source = SineWave::new(440.0).take_duration(Duration::from_secs_f32(0.25)).amplify(0.20);
+//! let source = SineWave::new(440.0)
+//!     .take_duration(Duration::from_secs_f32(0.25))
+//!     .amplify(Factor::Linear(0.20));
 //! player.append(source);
 //!
 //! // The sound plays in a separate thread. This call will block the current thread until the
@@ -230,6 +233,7 @@ pub mod decoder;
 pub mod const_source;
 pub mod fixed_source;
 
+pub mod effects;
 pub mod generators;
 
 pub mod math;